Try to fix gh-actions.
gh-actions tries to run py311-plone52, which is not wanted, and fails:
tox-gh-actions couldn't understand the parallel option. ignoring the given option: 0

There are multiple ways in which to solve this, I think.
After this commit, we would have two gh-actions jobs: py38 and py311.
The py38 one would run two tox envs in parallel: for 52 and 60.